+++ src/components/Engivia/Item/index.tsx
@@ -1,19 +1,14 @@
-import React from "react";
+import type { FC } from "react";
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
+import { EngiviaType } from "src/types/interface";
 
-export const Item = ({ children, ...rest }: any) => {
-  return (
-    <div className="p-2 text-sm bg-white rounded shadow-sm cursor-pointer">
-      {children}
-    </div>
-  );
+type Props = {
+  engivia: EngiviaType;
 };
 
-export function SortableItem({ id, children }: any) {
-  const sortable = useSortable({
-    id,
-  });
+export const SortableItem: FC<Props> = ({ engivia }) => {
+  const sortable = useSortable({ id: engivia.id });
   const { attributes, setNodeRef, listeners, transform, transition } = sortable;
 
   const style = {
@@ -24,7 +19,17 @@ export function SortableItem({ id, children }: any) {
 
   return (
     <div ref={setNodeRef} style={style} {...attributes} {...listeners}>
-      <Item>{children}</Item>
+      <div className="py-4 px-4 bg-white rounded-md shadow-sm">
+        <span className="text-lg">{engivia.body}</span>
+        <div className="flex items-center mt-4">
+          <img
+            className="mr-2 h-6 rounded-full"
+            src={engivia.postUser.image}
+            alt="avatar"
+          />
+          <span>{engivia.postUser.name}</span>
+        </div>
+      </div>
     </div>
   );
-}
+};
